refactor(frontend): data selection state
Group data selection state code in `src/state/data-selection`.
