+++ frontend/src/sidebar/networks/DamageSourceControl.tsx
@@ -19,7 +19,7 @@ import {
   damageSourceState,
   damageSourceStateEffect,
   damageTypeState,
-} from 'state/damage-mapping/damage-map';
+} from 'state/data-selection/damage-mapping/damage-map';
 import { HAZARDS_METADATA, HAZARDS_UI_ORDER } from 'config/view-layers/hazards/metadata';
 import { LayerStylePanel } from 'sidebar/ui/LayerStylePanel';
