+++ frontend/src/state/data-selection/damage-mapping/damage-map.ts
@@ -3,8 +3,8 @@ import { atom, selector } from 'recoil';
 
 import { HAZARD_DOMAINS } from 'config/view-layers/hazards/domains';
 import { dataParamOptionsState, dataParamState } from 'state/data-params';
-import { hazardSelectionState } from 'state/hazards/hazard-selection';
-import { networksStyleState } from 'state/networks/networks-style';
+import { hazardSelectionState } from 'state/data-selection/hazards/hazard-selection';
+import { networksStyleState } from 'state/data-selection/networks/networks-style';
 
 export const showDamagesState = selector({
   key: 'showDamagesState',
