+++ frontend/src/state/data-selection/drought/drought-parameters.ts
