+++ frontend/src/state/data-selection/hazards/hazard-selection.ts
