+++ frontend/src/state/data-selection/networks/network-selection.ts
