+++ frontend/src/state/data-selection/networks/networks-style.ts
@@ -1,5 +1,5 @@
 import { selector } from 'recoil';
-import { sectionStyleValueState } from '../sections';
+import { sectionStyleValueState } from 'state/sections';
 
 export const networksStyleState = selector({
   key: 'networksStyleState',
