+++ frontend/src/state/layers/view-layers/buildings.ts
@@ -3,7 +3,7 @@ import { selector } from 'recoil';
 import { ViewLayer } from 'lib/data-map/view-layers';
 import { truthyKeys } from 'lib/helpers';
 import { buildingsViewLayer } from 'config/view-layers/buildings/buildings-view-layer';
-import { buildingSelectionState } from 'state/buildings';
+import { buildingSelectionState } from 'state/data-selection/buildings';
 import { sectionVisibilityState } from 'state/sections';
 
 export const buildingsLayerState = selector<ViewLayer[]>({
