+++ frontend/src/state/layers/view-layers/droughtOptions.ts
@@ -7,7 +7,7 @@ import {
   droughtOptionsVariableState,
   droughtRcpParamState,
   droughtShowOptionsState,
-} from 'state/drought/drought-parameters';
+} from 'state/data-selection/drought/drought-parameters';
 import { sectionVisibilityState } from 'state/sections';
 
 export const droughtOptionsFieldSpecState = selector<FieldSpec>({
