+++ frontend/src/state/layers/view-layers/hazards.ts
@@ -4,7 +4,7 @@ import { ViewLayer } from 'lib/data-map/view-layers';
 import { truthyKeys } from 'lib/helpers';
 import { selector } from 'recoil';
 import { dataParamsByGroupState } from 'state/data-params';
-import { hazardVisibilityState } from 'state/hazards/hazard-visibility';
+import { hazardVisibilityState } from 'state/data-selection/hazards/hazard-visibility';
 import { sectionVisibilityState } from 'state/sections';
 
 export const hazardsLayerState = selector<ViewLayer[]>({
