+++ frontend/src/state/layers/view-layers/marine.ts
@@ -5,7 +5,7 @@ import { selector } from 'recoil';
 import { sectionStyleValueState, sectionVisibilityState } from 'state/sections';
 import { featureProperty } from 'lib/deck/props/data-source';
 import { Accessor } from 'lib/deck/props/getters';
-import { marineFiltersState } from 'state/solutions/marine-filters';
+import { marineFiltersState } from 'state/data-selection/solutions/marine-filters';
 
 export function habitatColorMap(x: string) {
   return MARINE_HABITAT_COLORS[x]?.css ?? MARINE_HABITAT_COLORS['other'].css;
